fix. view controller 분리
