+++ src/controller/cstation.js
@@ -19,7 +19,7 @@ export function clickAddStation() {
 
 export function constructStation() {
   const stationAddButton = document.getElementById(S.ADDBTN);
-  stationAddButton.addEventListener('click', clickAddStation);
+  stationAddButton.addEventListener('click', () => clickAddStation());
   document.getElementById(S.TABLE).addEventListener('click', (e) => {
     const target = e.target;
     const action = target.dataset.action;
